+++ lizard_progress/urls.py
@@ -95,6 +95,13 @@ urlpatterns = patterns(
     url('^projects/(?P<project_slug>[^/]+)/dashboard/$',
         login_required(DashboardView.as_view()),
         name='lizard_progress_dashboardview'),
+
+    url('^projects/(?P<project_slug>[^/]+)/(?P<contractor_slug>[^/]+)/planning/$',
+        login_required(views.PlanningView.as_view()),
+        name='lizard_progress_planningview'),
+    url('^projects/(?P<project_slug>[^/]+)/(?P<contractor_slug>[^/]+)/(?P<mtype_slug>[^/]+)/planning$',
+        login_required(views.PlanningView.as_view()),
+        name='lizard_progress_planningview'),
     url('^dashboardcsv/(?P<project_slug>[^/]+)/(?P<contractor_slug>[^/]+)/$',
         login_required(DashboardCsvView.as_view()),
         name='lizard_progress_dashboardcsvview'),